requirements.txt
streamlit
tensorflow
fastapi
uvicorn
boto3
pillow
requests
